Moved files and added logic controllers
- Moved loads of types out of huge files.
- Separated logic into dedicated logic controllers which can be easily tested in the future.
- Changed function names so that they make more sense, and split code up so that functions are more readable and logical.
