+++ Sources/Paywall/Models/Assignment/Assignment.swift
@@ -1,0 +1,13 @@
+//
+//  Assignment.swift
+//  Paywall
+//
+//  Created by Yusuf Tör on 02/03/2022.
+//
+
+import Foundation
+
+struct Assignment: Codable {
+  var experimentId: String
+  var variantId: String
+}
