+++ Sources/Paywall/Models/Assignment/ConfirmAssignmentResponse.swift
@@ -1,0 +1,12 @@
+//
+//  ConfirmAssignmentResponse.swift
+//  Paywall
+//
+//  Created by Yusuf Tör on 02/03/2022.
+//
+
+import Foundation
+
+struct ConfirmAssignmentResponse: Codable {
+  var status: String
+}
