+++ Sources/Paywall/Models/Assignment/ConfirmAssignments.swift
@@ -1,0 +1,12 @@
+//
+//  ConfirmAssignments.swift
+//  Paywall
+//
+//  Created by Yusuf Tör on 02/03/2022.
+//
+
+import Foundation
+
+struct ConfirmAssignments: Codable {
+  var assignments: [Assignment]
+}
