+++ Sources/Paywall/Models/EmptyResponse.swift
@@ -1,0 +1,10 @@
+//
+//  EmptyResponse.swift
+//  Paywall
+//
+//  Created by Yusuf Tör on 02/03/2022.
+//
+
+import Foundation
+
+struct EmptyResponse: Decodable {}
