+++ Sources/Paywall/Models/Events/EventsRequest.swift
@@ -1,0 +1,12 @@
+//
+//  EventsRequest.swift
+//  Paywall
+//
+//  Created by Yusuf Tör on 02/03/2022.
+//
+
+import Foundation
+
+struct EventsRequest: Codable {
+  var events: [JSON]
+}
