+++ Sources/Paywall/Models/Events/EventsResponse.swift
@@ -1,0 +1,12 @@
+//
+//  EventsResponse.swift
+//  Paywall
+//
+//  Created by Yusuf Tör on 02/03/2022.
+//
+
+import Foundation
+
+struct EventsResponse: Codable {
+  var status: String
+}
