+++ Sources/Paywall/Models/Paywall/PaywallPresentationStyle.swift
@@ -1,0 +1,14 @@
+//
+//  PaywallPresentationStyle.swift
+//  Paywall
+//
+//  Created by Yusuf Tör on 02/03/2022.
+//
+
+import Foundation
+
+enum PaywallPresentationStyle: String, Decodable {
+  case sheet = "SHEET"
+  case modal = "MODAL"
+  case fullscreen = "FULLSCREEN"
+}
