+++ Sources/Paywall/Models/Postback/Postback.swift
@@ -1,0 +1,25 @@
+//
+//  Postback.swift
+//  Paywall
+//
+//  Created by Yusuf Tör on 02/03/2022.
+//
+
+import Foundation
+import StoreKit
+
+struct Postback: Codable {
+  var products: [PostbackProduct]
+}
+
+struct PostbackProduct: Codable {
+  var identifier: String
+  var productVariables: JSON
+  var product: SWProduct
+
+  init(product: SKProduct) {
+    self.identifier = product.productIdentifier
+    self.productVariables = product.productVariables
+    self.product = SWProduct(product: product)
+  }
+}
