+++ Sources/Paywall/Models/Product/ProductVariable.swift
@@ -1,0 +1,13 @@
+//
+//  ProductVariable.swift
+//  Paywall
+//
+//  Created by Yusuf Tör on 02/03/2022.
+//
+
+import Foundation
+
+struct ProductVariable: Decodable {
+  var key: String
+  var value: JSON
+}
