+++ Sources/Paywall/Models/Product/Variable.swift
@@ -1,0 +1,13 @@
+//
+//  Variable.swift
+//  Paywall
+//
+//  Created by Yusuf Tör on 02/03/2022.
+//
+
+import Foundation
+
+struct Variable: Decodable {
+  var key: String
+  var value: JSON
+}
