+++ Sources/Paywall/Models/Template/TemplateDevice.swift
