+++ Sources/Paywall/Models/Triggers/TriggerInfo.swift
