--- Sources/Paywall/Types/MiscTypes.swift
@@ -1,10 +1,0 @@
-//
-//  File.swift
-//  
-//
-//  Created by Jake Mor on 12/26/21.
-//
-
-import Foundation
-
-struct EmptyResponse: Decodable {}
